+++ doc/Introduction-to-formulaic.R
@@ -1,16 +1,16 @@
-## ---- include = FALSE----------------------------------------------------
+## ---- include = FALSE---------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
 
-## ----setup, echo=FALSE---------------------------------------------------
+## ----setup, echo=FALSE--------------------------------------------------------
 library(formulaic)
 library(data.table)
 library(knitr)
 data("snack.dat", package = "formulaic")
 
-## ----constant,echo=FALSE-------------------------------------------------
+## ----constant,echo=FALSE------------------------------------------------------
 id.name <- "User ID"
 awareness.name <- "Awareness"
 satisfaction.name <- "Satisfaction"
@@ -34,39 +34,39 @@ include.backtick <- "as.needed"
 format.as <- "formula"
 force.main.effects <- TRUE
 
-## ----example-------------------------------------------------------------
+## ----example------------------------------------------------------------------
 awareness.name = "Awareness"
 variable.names = c("Age", "Gender", "Income Group", "Region", "Persona", "Typo")
 
 ex.form <-
   create.formula(outcome.name = awareness.name,
-                 input.names = variable.names,
+                 input.names = 'variable.names',
                  dat = snack.dat)
 
 ex.form$formula
 lm_example <- lm(formula = ex.form, data = snack.dat)
 summary(lm_example)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 user.outcome.name <- "Satisfaction"
 user.input.names <- c('Age Group', 'Gender', 'Region')
 
 create.formula(outcome.name = user.outcome.name, input.names = user.input.names)$formula
 
-## ----dataset-------------------------------------------------------------
+## ----dataset------------------------------------------------------------------
 list(dim(snack.dat), names(snack.dat))
 
-## ----add.backtick example------------------------------------------------
-as.needed = formulaic:::add.backtick(x = names(snack.dat), include.backtick = 'as.needed')
-all = formulaic:::add.backtick(x = names(snack.dat), include.backtick = 'all')
+## ----add.backtick example-----------------------------------------------------
+as.needed = formulaic:::add.backtick(x = names(snack.dat), i.backtick = 'as.needed')
+all = formulaic:::add.backtick(x = names(snack.dat), i.backtick = 'all')
 
 library(DT)
 datatable(data = cbind(as.needed, all))
 
-## ----add.backtick example 2----------------------------------------------
+## ----add.backtick example 2---------------------------------------------------
 create.formula(outcome.name = awareness.name, input.names = variable.names)$formula
 
-## ----add.backtick example 3----------------------------------------------
+## ----add.backtick example 3---------------------------------------------------
 create.formula(
   outcome.name = awareness.name,
   input.names = variable.names,
@@ -74,15 +74,15 @@ create.formula(
   include.backtick = "all"
 )$formula
 
-## ----add.backtick example 4----------------------------------------------
+## ----add.backtick example 4---------------------------------------------------
 create.formula(
   outcome.name = awareness.name,
   input.names = variable.names,
   format.as = "character",
-  include.backtick = "as.needed"
+  include.backtick = "all"
 )$formula
 
-## ----create.formula examples basic---------------------------------------
+## ----create.formula examples basic--------------------------------------------
 outcome.name.awareness <- "Awareness"
 input.names <-
   c("Age", "Gender", "Income", "Region", "Persona", "Typo")
@@ -94,7 +94,7 @@ basic.form <-
 
 print(basic.form)
 
-## ----create.formula example interactions---------------------------------
+## ----create.formula example interactions--------------------------------------
 interactions = list(c("Age Group", "Gender"),
                     c("Age Group", "Region"),
                     c("Age Group", "Gender", "Region"))
@@ -109,7 +109,7 @@ interaction.form <-
 
 print(interaction.form)
 
-## ----create.formula example input.patterns-------------------------------
+## ----create.formula example input.patterns------------------------------------
 bp.pattern = "BP_"
 input.patterns = c("Gend", bp.pattern)
 
@@ -123,7 +123,7 @@ pattern.form <-
 
 print(pattern.form)
 
-## ----create.formula example dot.1----------------------------------------
+## ----create.formula example dot.1---------------------------------------------
 dot.form.1 <-
   create.formula(outcome.name = outcome.name.awareness,
                  input.names = ".",
@@ -131,7 +131,7 @@ dot.form.1 <-
 
 print(dot.form.1)
 
-## ----create.formula example dot.2----------------------------------------
+## ----create.formula example dot.2---------------------------------------------
 
 input.names = c("Gender", ".")
 
@@ -139,7 +139,7 @@ dot.form.2 <- create.formula(outcome.name = outcome.name.awareness, input.names 
 
 print(dot.form.2)
 
-## ----create.formula example dot.3----------------------------------------
+## ----create.formula example dot.3---------------------------------------------
 
 input.names = c("Typo", ".")
 
@@ -147,7 +147,7 @@ dot.form.2 <- create.formula(outcome.name = outcome.name.awareness, input.names 
 
 print(dot.form.2)
 
-## ----create.formual example variables.to.exclude.form--------------------
+## ----create.formual example variables.to.exclude.form-------------------------
 input.names <-
   c("Age",
     "Gender",
@@ -180,7 +180,7 @@ variables.to.exclude.form <-
 
 print(variables.to.exclude.form)
 
-## ----create.formula outcomes as inputs-----------------------------------
+## ----create.formula outcomes as inputs----------------------------------------
 input.names <- c("Income", "Age", "Income")
 income.name = "Income"
 
@@ -191,7 +191,7 @@ outcomes.as.inputs.form <-
 
 print(outcomes.as.inputs.form)
 
-## ----create.formula example duplicated.inputs and interactions-----------
+## ----create.formula example duplicated.inputs and interactions----------------
 duplicated.inputs <- c(rep.int(x = "Age", times = 2), "Income")
 duplicated.interactions <-
   list(c("Age", "Income"), c("Age", "Income"))
@@ -206,7 +206,7 @@ duplicated.form <-
 
 print(duplicated.form)
 
-## ----create.formula example with typo------------------------------------
+## ----create.formula example with typo-----------------------------------------
 input.names <- c("Age", "Typo")
 income.name <- "Income"
 
@@ -214,17 +214,17 @@ formula.with.typo <-
   create.formula(outcome.name = income.name, input.names = input.names)
 print(formula.with.typo)
 
-## ----create.formula example without typo---------------------------------
+## ----create.formula example without typo--------------------------------------
 formula.without.typo <-
   create.formula(outcome.name = income.name,
                  input.names = input.names,
                  dat = snack.dat)
 print(formula.without.typo)
 
-## ----Numeric variable with no variation----------------------------------
+## ----Numeric variable with no variation---------------------------------------
 snack.dat[, .N, keyby = c("Awareness", "Consideration")]
 
-## ----lack of contrast example numerical variables------------------------
+## ----lack of contrast example numerical variables-----------------------------
 formula.consideration <-
   create.formula(outcome.name = consideration.name,
                  input.names = c(age.name, awareness.name))
@@ -235,7 +235,7 @@ glm(formula = formula.consideration,
     data = snack.dat,
     family = "binomial")$coefficients
 
-## ----create.formula with lack of contrast categorical 1------------------
+## ----create.formula with lack of contrast categorical 1-----------------------
 formula.consideration <-
   create.formula(
     outcome.name = consideration.name,
@@ -246,14 +246,14 @@ formula.consideration <-
 
 print(formula.consideration)
 
-## ----lack of contrast example categorical variables 0--------------------
+## ----lack of contrast example categorical variables 0-------------------------
 formula.awareness <-
   create.formula(outcome.name = awareness.name,
                  input.names = c(age.group.name, gender.name))
 
 print(formula.awareness$formula)
 
-## ----create.formula with lack of contrast 2, eval=FALSE, include=TRUE----
+## ----create.formula with lack of contrast 2, eval=FALSE, include=TRUE---------
 #  
 #  
 #  formula.awareness <-
@@ -266,7 +266,7 @@ print(formula.awareness$formula)
 #  
 #  print(formula.awareness)
 
-## ----create.fomula lack of contrast 3------------------------------------
+## ----create.fomula lack of contrast 3-----------------------------------------
 formula.consideration.1 <-
   create.formula(
     outcome.name = consideration.name,
@@ -278,7 +278,7 @@ formula.consideration.1 <-
 
 print(formula.consideration.1)
 
-## ----create.formula lack of contrast 4-----------------------------------
+## ----create.formula lack of contrast 4----------------------------------------
 formula.consideration.2 <-
   create.formula(
     outcome.name = consideration.name,
@@ -290,7 +290,7 @@ formula.consideration.2 <-
 
 print(formula.consideration.2)
 
-## ----create.formula large volume of categorical variables 01-------------
+## ----create.formula large volume of categorical variables 01------------------
 create.formula(
   outcome.name = satisfaction.name,
   input.names = c(age.name, income.name, region.name, id.name),
@@ -299,7 +299,7 @@ create.formula(
   max.input.categories = 30
 )$formula
 
-## ----create.formula large volume of categorical variables----------------
+## ----create.formula large volume of categorical variables---------------------
 create.formula(
   outcome.name = awareness.name,
   input.names = ".",
@@ -309,7 +309,7 @@ create.formula(
 )$formula
 
 
-## ----reduce.existing.formula example-------------------------------------
+## ----reduce.existing.formula example------------------------------------------
 the.initial.formula <- Awareness ~ .
 
 reduce.existing.formula(