+++ NEWS.md
@@ -1,14 +1,17 @@
+# formulaic 0.0.5
+* Create formula now supports 'all' in input names 
+* Added coverage test files
+* Update the Vignette
+* Refactor add.backtick function in Create formula to i.backtick  to address situtaion where  situal where include backtick  == 'all' outcome didn't properly compute. 
+
 # formulaic 0.0.4
 * Export function add.backtick 
 * Modified Formulaic Vignette
 
-
 # formulaic 0.0.3
-
 * now depends on R (>= 3.1.0). #feedback from reddit user 
 
 # formulaic 0.0.2
-
 * Added a `NEWS.md` file to track changes to the package.
 * Exported function add.backtick() as standalone function
 * Modified the wording of the package overview and function description